+++ RELEASENOTES.md
@@ -81,8 +81,8 @@
     *   Write `mfra` (Movie Fragment Random Access) box at the end of fragmented
         MP4 files to allow seeking via `FragmentedMp4Extractor` using
         `FLAG_READ_MFRA_FOR_SEEK_MAP`.
-    *   Fix crash in `Mp4Writer` when writing an EOS sample before any other
-        samples are written.
+    *   Fix crash in `Mp4Writer` and `FragmentedMp4Writer` when writing an EOS
+        sample before any other samples are written.
 *   IMA extension:
 *   Session:
     *   Add `MediaConstants.EXTRAS_KEY_PLAYLIST_ID` to populate